serialpy/requirements.txt
bidict==0.23.1
certifi==2024.12.14
charset-normalizer==3.4.1
h11==0.14.0
idna==3.10
joblib==1.4.2
numpy==2.2.1
pandas==2.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
python-engineio==4.11.2
python-socketio==5.12.1
pytz==2024.2
requests==2.32.3
scikit-learn==1.6.0
scipy==1.15.0
simple-websocket==1.1.0
six==1.17.0
threadpoolctl==3.5.0
tzdata==2024.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
